requirements.txt
Flask==3.0.0
pymongo==4.6.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
Authlib==1.3.0
requests==2.31.0

